+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## 1.6.4 under development
 
+- New #131: Add `multibyte` and `encoding` parameters to `Trim`, `LeftTrim`, `RightTrim` and `ToArrayOfStrings`
+  attributes and their resolvers (@vjik)
+- Chg #131: Deprecate the `..` range syntax in the `characters` parameter of `Trim`, `LeftTrim` and `RightTrim`
+  attributes (@vjik)
 - Enh #117: Explicitly import functions and constants in "use" section (@mspirkov)
 - Enh #132: Exclude development files from the distribution archive (@vjik)